Apply newfound knowledge
Apparently, the wiiu doesn't check for the truevision footer, but rather whether the file is long enough that it could contain a truevision footer.

Therefore, to also cover cases of the extension and developer area being used, "fixup" will no longer overwrite these if the file is found to be long enough to be accepted, no matter what it contains, and the check will also pass in the same cases.